+++ src/Arguments/ServiceBus/Queue/QueuePeekArguments.cs
@@ -1,0 +1,12 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+namespace AzureMcp.Arguments.ServiceBus.Queue;
+
+public class QueuePeekArguments : BaseQueueArguments
+{
+    /// <summary>
+    /// Maximum number of messages to peek from queue.
+    /// </summary>
+    public int? MaxMessages { get; set; }
+}