+++ src/Arguments/ServiceBus/Topic/BaseTopicArguments.cs
@@ -1,0 +1,22 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+namespace AzureMcp.Arguments.ServiceBus.Subscription;
+
+public class BaseTopicArguments : SubscriptionArguments
+{
+    /// <summary>
+    /// Service Bus namespace.
+    /// </summary>
+    public string? Namespace { get; set; }
+
+    /// <summary>
+    /// Name of the topic.
+    /// </summary>
+    public string? TopicName { get; set; }
+
+    /// <summary>
+    /// Name of the subscription.
+    /// </summary>
+    public string? SubscriptionName { get; set; }
+}