+++ src/Arguments/ServiceBus/Topic/SubscriptionPeekArguments.cs
@@ -1,0 +1,12 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+namespace AzureMcp.Arguments.ServiceBus.Subscription;
+
+public class SubscriptionPeekArguments : BaseTopicArguments
+{
+    /// <summary>
+    /// Maximum number of messages to peek from subscription.
+    /// </summary>
+    public int? MaxMessages { get; set; }
+}