+++ src/Commands/ServiceBus/ServiceBusJsonContext.cs
@@ -1,0 +1,15 @@
+﻿// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+using System.Text.Json.Serialization;
+using AzureMcp.Commands.ServiceBus.Queue;
+using AzureMcp.Commands.ServiceBus.Topic;
+
+namespace AzureMcp.Commands.ServiceBus;
+
+[JsonSerializable(typeof(QueuePeekCommand.QueuePeekCommandResult))]
+[JsonSerializable(typeof(SubscriptionPeekCommand.SubscriptionPeekCommandResult))]
+[JsonSourceGenerationOptions(PropertyNamingPolicy = JsonKnownNamingPolicy.CamelCase)]
+internal sealed partial class ServiceBusJsonContext : JsonSerializerContext
+{
+}